--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,5 +18,11 @@ thiserror = "1.0.30"
 wyhash = "0.5.0"
 
 [dev-dependencies]
+criterion = { version = "0.3.5", features = ["html_reports"] }
+pprof = { version = "0.6.2", features = ["flamegraph", "protobuf", "criterion"] }
 proptest = "1.0.0"
 rand = "0.8.5"
+
+[[bench]]
+name = "synchronizer"
+harness = false