--- a/src/synchronizer.rs
+++ b/src/synchronizer.rs
@@ -8,10 +8,10 @@ use std::hash::{BuildHasher, BuildHasherDefault, Hasher};
 use std::time::Duration;
 
 use bytecheck::CheckBytes;
-use rkyv::ser::serializers::AllocSerializer;
+use rkyv::ser::serializers::{AlignedSerializer, AllocSerializer};
 use rkyv::ser::Serializer;
 use rkyv::validation::validators::DefaultValidator;
-use rkyv::{archived_root, check_archived_root, Archive, Serialize};
+use rkyv::{archived_root, check_archived_root, AlignedVec, Archive, Serialize};
 use thiserror::Error;
 use wyhash::WyHash;
 
@@ -24,13 +24,15 @@ use crate::synchronizer::SynchronizerError::*;
 /// `Synchronizer` is a concurrency primitive that manages data access between a single writer process and multiple reader processes.
 ///
 /// It coordinates the access to two data files that store the shared data. A state file, also memory-mapped, stores the index of the current data file and the number of active readers for each index, updated via atomic instructions.
-pub struct Synchronizer<H: Hasher + Default = WyHash> {
+pub struct Synchronizer<H: Hasher + Default = WyHash, const N: usize = 1024> {
     /// Container storing state mmap
     state_container: StateContainer,
     /// Container storing data mmap
     data_container: DataContainer,
     /// Hasher used for checksum calculation
     build_hasher: BuildHasherDefault<H>,
+    /// Re-usable buffer for serialization
+    serialize_buffer: Option<AlignedVec>,
 }
 
 /// `SynchronizerError` enumerates all possible errors returned by this library.
@@ -61,16 +63,21 @@ pub enum SynchronizerError {
     InvalidInstanceVersionParams,
 }
 
-/// Default serializer with 1 MB scratch space allocated on the heap.
-type DefaultSerializer = AllocSerializer<1_000_000>;
-
 impl Synchronizer {
-    /// Create new instance of `Synchronizer` using given `path_prefix`
-    pub fn new(path_prefix: &OsStr) -> Synchronizer {
+    /// Create new instance of `Synchronizer` using given `path_prefix` and default template parameters
+    pub fn new(path_prefix: &OsStr) -> Self {
+        Self::with_params(path_prefix)
+    }
+}
+
+impl<H: Hasher + Default, const N: usize> Synchronizer<H, N> {
+    /// Create new instance of `Synchronizer` using given `path_prefix` and template parameters
+    pub fn with_params(path_prefix: &OsStr) -> Self {
         Synchronizer {
             state_container: StateContainer::new(path_prefix),
             data_container: DataContainer::new(path_prefix),
             build_hasher: BuildHasherDefault::default(),
+            serialize_buffer: Some(AlignedVec::new()),
         }
     }
 
@@ -97,11 +104,18 @@ impl Synchronizer {
         grace_duration: Duration,
     ) -> Result<(usize, bool), SynchronizerError>
     where
-        T: Serialize<DefaultSerializer>,
+        T: Serialize<AllocSerializer<N>>,
         T::Archived: for<'b> CheckBytes<DefaultValidator<'b>>,
     {
+        let mut buf = self.serialize_buffer.take().ok_or(FailedEntityWrite)?;
+        buf.clear();
+
         // serialize given entity into bytes
-        let mut serializer = DefaultSerializer::default();
+        let mut serializer = AllocSerializer::new(
+            AlignedSerializer::new(buf),
+            Default::default(),
+            Default::default(),
+        );
         let _ = serializer
             .serialize_value(entity)
             .map_err(|_| FailedEntityWrite)?;
@@ -126,6 +140,9 @@ impl Synchronizer {
         // switch readers to new version
         state.switch_version(new_version);
 
+        // Restore buffer for potential reuse
+        self.serialize_buffer.replace(data);
+
         Ok((size, reset))
     }
 
@@ -138,7 +155,7 @@ impl Synchronizer {
         grace_duration: Duration,
     ) -> Result<(usize, bool), SynchronizerError>
     where
-        T: Serialize<DefaultSerializer>,
+        T: Serialize<AllocSerializer<N>>,
         T::Archived: for<'b> CheckBytes<DefaultValidator<'b>>,
     {
         // fetch current state from mapped memory