List of changes:
* Added hasher template parameter `H` for `Synchronizer` with `WyHash` as default. This will speed up checksum calculation during writes.
* Added read/write benchmarks from `Synchronizer`.
* Optimize `Synchronizer::write` and `write_raw` methods to re-use opened maps.
* Speed up writes further with re-usable buffer backed by `AlignedVec`.

Benchmarks:
```
synchronizer/write      time:   [250.71 ns 251.42 ns 252.41 ns]
                        thrpt:  [3.9619 Melem/s 3.9774 Melem/s 3.9887 Melem/s]
                 change:
                        time:   [-99.152% -99.147% -99.142%] (p = 0.00 < 0.05)
                        thrpt:  [+11559% +11627% +11699%]
                        Performance has improved.
synchronizer/write_raw  time:   [145.25 ns 145.53 ns 145.92 ns]
                        thrpt:  [6.8531 Melem/s 6.8717 Melem/s 6.8849 Melem/s]
                 change:
                        time:   [-98.508% -98.471% -98.443%] (p = 0.00 < 0.05)
                        thrpt:  [+6322.7% +6441.2% +6602.9%]
                        Performance has improved.
synchronizer/read/check_bytes_true
                        time:   [40.114 ns 40.139 ns 40.186 ns]
                        thrpt:  [24.884 Melem/s 24.914 Melem/s 24.929 Melem/s]
                 change:
                        time:   [-0.1031% -0.0200% +0.0753%] (p = 0.69 > 0.05)
                        thrpt:  [-0.0752% +0.0200% +0.1032%]
                        No change in performance detected.
synchronizer/read/check_bytes_false
                        time:   [26.658 ns 26.673 ns 26.696 ns]
                        thrpt:  [37.458 Melem/s 37.491 Melem/s 37.512 Melem/s]
                 change:
                        time:   [-0.9845% -0.9083% -0.7959%] (p = 0.00 < 0.05)
                        thrpt:  [+0.8023% +0.9167% +0.9943%]
                        Change within noise threshold.
```